+++ selfdrive/car/ford/radar_interface.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-import numpy as np
+from collections import namedtuple
 from selfdrive.can.parser import CANParser
 from selfdrive.car.ford.values import DBC
 from cereal import car
@@ -8,47 +8,22 @@ import zmq
 from selfdrive.services import service_list
 from selfdrive.config import Conversions as CV
 import selfdrive.messaging as messaging
-from collections import namedtuple
+import selfdrive.rcp_helpers as rcp
 
-def create_radar_signals(*signals):
-  # accepts multiple namedtuples in the form ([('name', value)],[msg])
-  name_value = []
-  msgs = []
-  repetitions = []
-  for s in signals:
-    name_value += [nv for nv in s.name_value]
-    name_value_n = len(s.name_value)
-    msgs_n = [len(s.msg)]
-    repetitions += msgs_n * name_value_n
-    msgs += s.msg * name_value_n
-
-  name_value = sum([[nv] * r for nv, r in zip(name_value, repetitions)], [])
-  names = [n for n, v in name_value]
-  vals  = [v for n, v in name_value]
-  return zip(names, msgs, vals)
-
-def create_radar_checks(msgs, select, rate = [20]):
-  if select == "all":
-    return zip(msgs, rate * len(msgs))
-  if select == "last":
-    return zip([msgs[-1]], rate)
-  if select == "none":
-    return []
-  return []
 
 RADAR_MSGS = range(0x500, 0x540)
 
 def _create_radard_can_parser(car_fingerprint):
   dbc_f = DBC[car_fingerprint]['radar']
 
-  sig = namedtuple('sig','name_value msg')
+  sig = namedtuple('sig', 'name_value msg')
   trackers = [('X_Rel', 0),
               ('Angle', 0),
               ('V_Rel', 0)]
   tracker_sig = sig(trackers, RADAR_MSGS)
 
-  signals = create_radar_signals(tracker_sig)
-  checks = create_radar_checks(RADAR_MSGS, select = "all")
+  signals = rcp.create_radar_signals(tracker_sig)
+  checks = rcp.create_radar_checks(RADAR_MSGS, select = "all")
 
   return CANParser(dbc_f, signals, checks, 1)
